Spontaneous commit To-do: Get hitboxes working
