+++ Assets/Scripts/Attack.cs
@@ -1,0 +1,6 @@
+﻿using UnityEngine;
+using System.Collections;
+
+public struct Attack {
+    AttackData data;
+}
