+++ Assets/Scripts/AttackData.cs.meta
@@ -1,6 +1,6 @@
 fileFormatVersion: 2
-guid: bf6ac62a61c77b140b55068af9461fa4
-timeCreated: 1451932021
+guid: bf6690cf35e04a14580fa72e21f7a6e0
+timeCreated: 1454951360
 licenseType: Free
 MonoImporter:
   serializedVersion: 2
