+++ Assets/Scripts/FighterHealth.cs
@@ -1,0 +1,15 @@
+﻿using UnityEngine;
+using System.Collections;
+
+public class FighterHealth : MonoBehaviour {
+
+    // Use this for initialization
+    void Start () {
+
+    }
+
+    // Update is called once per frame
+    void Update () {
+
+    }
+}
