--- Assets/Scripts/Framework/FighterHealth.cs
@@ -1,15 +1,0 @@
-﻿using UnityEngine;
-using System.Collections;
-
-public class FighterHealth : MonoBehaviour {
-
-    // Use this for initialization
-    void Start () {
-
-    }
-
-    // Update is called once per frame
-    void Update () {
-
-    }
-}
