+++ Assets/Scripts/Movement/Fighter.cs.meta
@@ -1,0 +1,12 @@
+fileFormatVersion: 2
+guid: 199951c5eb210e64a92326de242a7df7
+timeCreated: 1446745575
+licenseType: Free
+MonoImporter:
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
