+++ Assets/Scripts/Movement/TestMover.cs.meta
@@ -1,0 +1,12 @@
+fileFormatVersion: 2
+guid: bf6ac62a61c77b140b55068af9461fa4
+timeCreated: 1451932021
+licenseType: Free
+MonoImporter:
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
