--- Assets/Scripts/SimpleRaycastEngine.cs.meta
@@ -1,12 +1,0 @@
-fileFormatVersion: 2
-guid: a432eb3ac9f55e14db1799995983f46a
-timeCreated: 1453830022
-licenseType: Free
-MonoImporter:
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
